requirements.txt
pandas>=2.0.3
numpy>=1.24.3
scikit-learn==1.3.0
pyyaml==6.0.1
SQLAlchemy==2.0.19
requests==2.31.0
APScheduler==3.10.1
matplotlib==3.7.2
seaborn==0.12.2
